requirements.txt
appdirs==1.4.4
APScheduler==3.9.1
arabic-reshaper==2.1.3
asgiref==3.5.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
black==22.6.0
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
cssselect2==0.6.0
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0.4
django-apscheduler==0.6.2
django-cors-headers==3.11.0
django-filter==21.1
django-picklefield==3.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-recursive==0.1.2
drf-nested-routers==0.93.4
drf-spectacular==0.22.1
et-xmlfile==1.1.0
fonttools==4.34.4
future==0.18.3
getmac==0.9.1
ghp-import==2.1.0
gunicorn==20.1.0
html5lib==1.1
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.4.0
lxml==4.9.1
Markdown==3.3.6
MarkupSafe==2.1.1
mergedeep==1.3.4
mkdocs==1.4.2
mypy-extensions==0.4.3
numpy==1.23.0
oauthlib==3.2.2
openpyxl==3.0.10
oscrypto==1.3.0
packaging==21.3
pandas==1.4.3
pathspec==0.9.0
pdfkit==1.0.0
Pillow==9.2.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.9.1
pydyf==0.2.0
pyee==8.2.2
pyHanko==0.13.2
pyhanko-certvalidator==0.19.5
PyJWT==2.3.0
pyparsing==3.0.9
PyPDF3==1.0.6
pyphen==0.12.0
pyppeteer==1.0.2
pyrsistent==0.18.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyyaml_env_tag==0.1
qrcode==7.3.1
redis==4.3.3
regex==2022.3.2
reportlab==3.6.10
requests==2.27.1
requests-futures==1.0.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
svglib==1.3.0
tinycss2==1.1.1
tomli==2.0.1
tqdm==4.64.0
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.26.9
watchdog==2.2.1
weasyprint==56.0
webencodings==0.5.1
websockets==10.4
whitenoise==6.0.0
wrapt==1.14.1
XlsxWriter==3.0.3
xlwt==1.3.0
zipp==3.8.0
zopfli==0.2.1
